Move account/accountdatabase to appropriate folders
